feat: add component productlist
